+++ default_planner/scheduler.cpp
@@ -3,12 +3,15 @@
 namespace DefaultPlanner{
 
 std::mt19937 mt;
+std::unordered_set<int> free_agents;
+std::unordered_set<int> free_tasks;
 
 void schedule_initialize(int preprocess_time_limit, SharedEnvironment* env)
 {
     // cout<<"schedule initialise limit" << preprocess_time_limit<<endl;
     DefaultPlanner::init_heuristics(env);
     mt.seed(0);
+    return;
 }
 
 void schedule_plan(int time_limit, std::vector<int> & proposed_schedule,  SharedEnvironment* env)
@@ -18,55 +21,73 @@ void schedule_plan(int time_limit, std::vector<int> & proposed_schedule,  Shared
     TimePoint endtime = std::chrono::steady_clock::now() + std::chrono::milliseconds(time_limit);
     // cout<<"schedule plan limit" << time_limit <<endl;
 
-    proposed_schedule.resize(env->num_of_agents, -1); //default no schedule
+    // the default scheduler keep track of all the free agents and unassigned (=free) tasks across timesteps
+    free_agents.insert(env->new_freeagents.begin(), env->new_freeagents.end());
+    free_tasks.insert(env->new_tasks.begin(), env->new_tasks.end());
 
-
-    int i_task, min_task_i, min_task_makespan, dist, c_loc, count;
+    int min_task_i, min_task_makespan, dist, c_loc, count;
     clock_t start = clock();
 
-    for (int i = 0; i < env->num_of_agents && std::chrono::steady_clock::now() < endtime; i++)
+    // iterate over the free agents to decide which task to assign to each of them
+    std::unordered_set<int>::iterator it = free_agents.begin();
+    while (it != free_agents.end())
     {
-        
-        if (env->curr_task_schedule[i] == -1)
+        //keep assigning until timeout
+        if (std::chrono::steady_clock::now() > endtime)
         {
-            
-            min_task_i = -1;
-            min_task_makespan = INT_MAX;
-            count = 0;
-            for (i_task=0 ; i_task < env->task_pool.size() && std::chrono::steady_clock::now() < endtime ;i_task++)
-            {                
-
-                if (env->task_pool[i_task].agent_assigned != -1)
-                    continue;
-                dist = 0;
-                c_loc = env->curr_states.at(i).location;
-                for (int loc : env->task_pool[i_task].locations){
-                    dist += DefaultPlanner::get_h(env, c_loc, loc);
-                    c_loc = loc;
-                }
-                if (dist < min_task_makespan){
-                    min_task_i = i_task;
-                    min_task_makespan = dist;
-                }
-                count++;            
-            }
-
-
-            if (min_task_i != -1){
-                proposed_schedule[i] = env->task_pool[min_task_i].task_id;
-                env->task_pool[min_task_i].agent_assigned = i;
-            }
-            else{
-                proposed_schedule[i] = -1;
-            }
-            
-
+            break;
         }
-        else
+        int i = *it;
+
+        assert(env->curr_task_schedule[i] == -1);
+            
+        min_task_i = -1;
+        min_task_makespan = INT_MAX;
+        count = 0;
+
+        // iterate over all the unassigned tasks to find the one with the minimum makespan for agent i
+        for (int t_id : free_tasks)
         {
-            proposed_schedule[i] = env->curr_task_schedule[i];
+            //check for timeout every 10 task evaluations
+            if (count % 10 == 0 && std::chrono::steady_clock::now() > endtime)
+            {
+                break;
+            }
+            dist = 0;
+            c_loc = env->curr_states.at(i).location;
+
+            // iterate over the locations (errands) of the task to compute the makespan to finish the task
+            // makespan: the time for the agent to complete all the errands of the task t_id in order
+            for (int loc : env->task_pool[t_id].locations){
+                dist += DefaultPlanner::get_h(env, c_loc, loc);
+                c_loc = loc;
+            }
+
+            // update the new minimum makespan
+            if (dist < min_task_makespan){
+                min_task_i = t_id;
+                min_task_makespan = dist;
+            }
+            count++;            
+        }
+
+        // assign the best free task to the agent i (assuming one exists)
+        if (min_task_i != -1){
+            proposed_schedule[i] = min_task_i;
+            it = free_agents.erase(it);
+            free_tasks.erase(min_task_i);
+        }
+        // nothing to assign
+        else{
+            proposed_schedule[i] = -1;
+            it++;
         }
     }
-    // cout << ((float)(clock() - start))/CLOCKS_PER_SEC <<endl;
+    #ifndef NDEBUG
+    cout << "Time Usage: " <<  ((float)(clock() - start))/CLOCKS_PER_SEC <<endl;
+    cout << "new free agents: " << env->new_freeagents.size() << " new tasks: "<< env->new_tasks.size() <<  endl;
+    cout << "free agents: " << free_agents.size() << " free tasks: " << free_tasks.size() << endl;
+    #endif
+    return;
 }
 }