+++ inc/CompetitionSystem.h
@@ -22,7 +22,6 @@ public:
     {
         num_of_agents = start_locs.size();
         starts.resize(num_of_agents);
-        paths.resize(num_of_agents);
 
         for (size_t i = 0; i < start_locs.size(); i++)
             {
@@ -65,14 +64,15 @@ public:
     void set_num_tasks_reveal(float num){task_manager.set_num_tasks_reveal(num);};
     void set_plan_time_limit(int limit){plan_time_limit = limit;};
     void set_preprocess_time_limit(int limit){preprocess_time_limit = limit;};
+    void set_log_level(int level){log_level = level;};
     void set_logger(Logger* logger){
         this->logger = logger;
         task_manager.set_logger(logger);
     }
 
     void simulate(int simulation_time);
-    void plan(vector<Action> & actions,vector<int> & proposed_schedule);
-    std::pair<vector<Action> ,vector<int>> plan_wrapper();
+    void plan(int & timeout_timesteps);
+    bool planner_wrapper();
 
     //void saveSimulationIssues(const string &fileName) const;
     void saveResults(const string &fileName, int screen) const;
@@ -80,32 +80,30 @@ public:
 
 protected:
     Grid map;
+    int simulation_time;
 
-    std::future<std::pair<vector<Action> ,vector<int> >> future;
+    vector<Action> proposed_actions;
+    vector<int> proposed_schedule;
+
+    int total_timetous = 0;
+
+
+    std::future<bool> future;
     std::thread task_td;
     bool started = false;
 
     Entry* planner;
     SharedEnvironment* env;
 
-    //ActionModelWithRotate* model;
-
-    // #timesteps for simulation
-    //int timestep;
-
     int preprocess_time_limit=10;
 
     int plan_time_limit = 3;
 
-    std::vector<Path> paths;
 
     vector<State> starts;
     int num_of_agents;
 
-    //vector<State> curr_states;
-
-    vector<list<Action>> actual_movements;
-    vector<list<Action>> planner_movements;
+    int log_level = 1;
 
     // tasks that haven't been finished but have been revealed to agents;