+++ inc/TaskManager.h
@@ -14,6 +14,9 @@ public:
     vector<list<pair<int,int>>> planner_schedule;
     list<std::tuple<std::string,int,int,int,int>> schedule_errors;
 
+    vector<int> new_freeagents;
+    vector<int> new_tasks;
+
     list<int> check_finished_tasks(vector<State> states, int timestep);
 
     int curr_timestep;
@@ -30,6 +33,8 @@ public:
 
     bool validate_task_assgnment(vector<int> assignment); // validate the task assignment
     bool set_task_assignment( vector<int>  assignment); // set the task assignment; return true if task is valid
+
+    int get_number_errors() const {return schedule_errors.size();}
 
 
 
@@ -61,7 +66,7 @@ public:
 private:
     Logger* logger = nullptr;
 
-    unordered_map<int, Task*> ongoing_tasks;
+    std::unordered_map<int, Task*> ongoing_tasks;
     vector<int> current_assignment;
 
     int num_tasks_reveal = 1;