+++ python/pyMAPFPlanner.py
@@ -12,8 +12,6 @@ class pyMAPFPlanner:
         if env is not None:
             self.env = env
 
-        #print("pyMAPFPlanner created!  python debug")
-
     def initialize(self, preprocess_time_limit: int):
         """_summary_
 
@@ -21,10 +19,7 @@ class pyMAPFPlanner:
             preprocess_time_limit (_type_): _description_
         """
         pass
-        # testlib.test_torch()
-        #print("planner initialize done... python debug")
         return True
-        # raise NotImplementedError()
 
     def plan(self, time_limit):
         """_summary_
@@ -47,25 +42,17 @@ class pyMAPFPlanner:
 
         # example of only using single-agent search
         return self.sample_priority_planner(int(time_remaining.total_seconds() * 1000))
-        # #print("python binding debug")
-        # #print("env.rows=",self.env.rows,"env.cols=",self.env.cols,"env.map=",self.env.map)
-        # raise NotImplementedError("YOU NEED TO IMPLEMENT THE PYMAPFPLANNER!")
 
     def naive_a_star(self,time_limit):
-        #print("I am planning")
         actions = [MAPF.Action.W for i in range(len(self.env.curr_states))]
         for i in range(0, self.env.num_of_agents):
-            #print("python start plan for agent ", i, end=" ")
             path = []
             if len(self.env.goal_locations[i]) == 0:
-                #print(i, " does not have any goal left", end=" ")
                 path.append((self.env.curr_states[i].location, self.env.curr_states[i].orientation))
             else:
-                #print(" with start and goal: ", end=" ")
                 path = self.single_agent_plan(
                     self.env.curr_states[i].location, self.env.curr_states[i].orientation, self.env.goal_locations[i][0][0])
 
-            #print("current location:", path[0][0],"current direction: ", path[0][1])
             if path[0][0] != self.env.curr_states[i].location:
                 actions[i] = MAPF.Action.FW
             elif path[0][1] != self.env.curr_states[i].orientation:
@@ -74,15 +61,11 @@ class pyMAPFPlanner:
                     actions[i] = MAPF.Action.CR
                 elif incr == -1 or incr == 3:
                     actions[i] = MAPF.Action.CCR
-        # #print(actions)
         actions = [int(a) for a in actions]
-        # #print(actions)
         return np.array(actions, dtype=int)
 
     def single_agent_plan(self, start: int, start_direct: int, end: int):
-        #print(start, start_direct, end)
         path = []
-        # AStarNode (u,dir,t,f)
         open_list = PriorityQueue()
         s = (start, start_direct, 0, self.getManhattanDistance(start, end))
         open_list.put([0, s])
@@ -103,7 +86,6 @@ class pyMAPFPlanner:
 
                 break
             neighbors = self.getNeighbors(curr[0], curr[1])
-            # #print("neighbors=",neighbors)
             for neighbor in neighbors:
                 if (neighbor[0]*4+neighbor[1]) in close_list:
                     continue
@@ -111,7 +93,6 @@ class pyMAPFPlanner:
                              self.getManhattanDistance(neighbor[0], end))
                 parent[(next_node[0], next_node[1])] = (curr[0], curr[1])
                 open_list.put([next_node[3]+next_node[2], next_node])
-        #print(path)
         return path
 
     def getManhattanDistance(self, loc1: int, loc2: int) -> int:
@@ -151,23 +132,19 @@ class pyMAPFPlanner:
         if (new_direction == 4):
             new_direction = 0
         neighbors.append((location, new_direction))
-        # #print("debug!!!!!!!", neighbors)
         return neighbors
 
     def space_time_plan(self,start: int, start_direct: int, end: int, reservation: Set[Tuple[int, int, int]]) -> List[Tuple[int, int]]:
-        #print(start, start_direct, end)
         path = []
         open_list = PriorityQueue()
         all_nodes = {}  # loc+dict, t
         parent={}
         s = (start, start_direct, 0, self.getManhattanDistance(start, end))
         open_list.put((s[3], id(s), s))
-        # all_nodes[(start * 4 + start_direct, 0)] = s
         parent[(start * 4 + start_direct, 0)]=None
 
         while not open_list.empty():
             n=open_list.get()
-            # #print("n=",n)
             _, _, curr = n
         
             curr_location, curr_direction, curr_g, _ = curr
@@ -181,7 +158,6 @@ class pyMAPFPlanner:
                     curr=parent[(curr[0]*4+curr[1],curr[2])]
                     if curr is None:
                         break
-                    # curr = curr[5]
                 path.pop()
                 path.reverse()
                 break
@@ -213,10 +189,6 @@ class pyMAPFPlanner:
                 
                     parent[(neighbor_location * 4 +
                             neighbor_direction, next_node[2])]=curr
-
-        # for v in path:
-        #     print(f"({v[0]},{v[1]}), ", end="")
-        # print()
         return path
 
     def sample_priority_planner(self,time_limit:int):
@@ -224,18 +196,14 @@ class pyMAPFPlanner:
         reservation = set()  # loc1, loc2, t
 
         for i in range(self.env.num_of_agents):
-            #print("start plan for agent", i)
             path = []
             if not self.env.goal_locations[i]:
-                #print(", which does not have any goal left.")
                 path.append((self.env.curr_states[i].location, self.env.curr_states[i].orientation))
                 reservation.add((self.env.curr_states[i].location, -1, 1))
 
         for i in range(self.env.num_of_agents):
-            #print("start plan for agent", i)
             path = []
             if self.env.goal_locations[i]:
-                #print("with start and goal:")
                 path = self.space_time_plan(
                     self.env.curr_states[i].location,
                     self.env.curr_states[i].orientation,
@@ -244,7 +212,6 @@ class pyMAPFPlanner:
                 )
             
             if path:
-                #print("current location:", path[0][0], "current direction:", path[0][1])
                 if path[0][0] != self.env.curr_states[i].location:
                     actions[i] = MAPF.Action.FW
                 elif path[0][1] != self.env.curr_states[i].orientation: