+++ src/CompetitionSystem.cpp
@@ -9,19 +9,6 @@
 using json = nlohmann::ordered_json;
 
 
-void BaseSystem::move(vector<Action>& actions)
-{
-
-    vector<State> curr_states = simulator.move(actions);
-    int timestep = simulator.get_curr_timestep();
-    // agents do not move
-
-    for (int k = 0; k < num_of_agents; k++)
-    {
-        paths[k].push_back(curr_states[k]);
-        actual_movements[k].push_back(actions[k]);
-    }
-}
 
 
 
@@ -33,77 +20,77 @@ void BaseSystem::move(vector<Action>& actions)
 // }
 
 
-void BaseSystem::sync_shared_env() {
+void BaseSystem::sync_shared_env() 
+{
+    if (!started)
+    {
+        env->goal_locations.resize(num_of_agents);
+        task_manager.sync_shared_env(env);
+        simulator.sync_shared_env(env);
 
-  if (!started){
-      env->goal_locations.resize(num_of_agents);
-      task_manager.sync_shared_env(env);
-      simulator.sync_shared_env(env);
-  }
-  else
-  {
-    env->curr_timestep = simulator.get_curr_timestep();
-  }
+        if (simulator.get_curr_timestep() == 0)
+        {
+            env->new_freeagents.reserve(num_of_agents); //new free agents are empty in task_manager on initialization, set it after task_manager sync
+            for (int i = 0; i < num_of_agents; i++)
+            {
+                env->new_freeagents.push_back(i);
+            }
+        }
+        //update proposed action to all wait
+        proposed_actions.clear();
+        proposed_actions.resize(num_of_agents, Action::W);
+        //update proposed schedule to previous assignment
+        proposed_schedule = env->curr_task_schedule;
+        
+    }
+    else
+    {
+        env->curr_timestep = simulator.get_curr_timestep();
+    }
 }
 
 
-std::pair<vector<Action> ,vector<int> > BaseSystem::plan_wrapper()
+bool BaseSystem::planner_wrapper()
 {
-    vector<Action> actions;
-    vector<int> proposed_schedule;
-    planner->compute(plan_time_limit, actions,proposed_schedule);
-    return {actions, proposed_schedule};
+    planner->compute(plan_time_limit, proposed_actions, proposed_schedule);
+    return true;
 }
 
 
-void BaseSystem::plan(vector<Action> & actions,vector<int> & proposed_schedule)
+void BaseSystem::plan(int & timeout_timesteps)
 {
-
-    int timestep = simulator.get_curr_timestep();
 
     using namespace std::placeholders;
-    if (started && future.wait_for(std::chrono::milliseconds(0)) != std::future_status::ready)
-    {
-        std::cout << started << "     " << (future.wait_for(std::chrono::milliseconds(0)) != std::future_status::ready) << std::endl;
-        if(logger)
-        {
-            logger->log_info("planner cannot run because the previous run is still running", timestep);
-        }
+    int timestep = simulator.get_curr_timestep();
 
-        if (future.wait_for(std::chrono::milliseconds(plan_time_limit)) == std::future_status::ready)
-        {
-            task_td.join();
-            started = false;
-            auto res = future.get();
-            actions = res.first;
-            proposed_schedule = res.second;
-            return;
-        }
-        logger->log_info("planner timeout", timestep);
-        return;
-    }
+    std::packaged_task<bool()> task(std::bind(&BaseSystem::planner_wrapper, this));
 
-    std::packaged_task<std::pair<vector<Action> ,vector<int> >()> task(std::bind(&BaseSystem::plan_wrapper, this));
+
     future = task.get_future();
-    if (task_td.joinable())
-    {
+    if (task_td.joinable()){
         task_td.join();
     }
     env->plan_start_time = std::chrono::steady_clock::now();
     task_td = std::thread(std::move(task));
-    started = true;
-    if (future.wait_for(std::chrono::milliseconds(plan_time_limit)) == std::future_status::ready)
-    {
-        task_td.join();
-        started = false;
-        auto res = future.get();
-        actions = res.first;
-        proposed_schedule = res.second;
 
-        return;
+    started = true;
+
+    while (timestep + timeout_timesteps < simulation_time){
+
+        if (future.wait_for(std::chrono::milliseconds(plan_time_limit)) == std::future_status::ready)
+            {
+                task_td.join();
+                started = false;
+                auto res = future.get();
+
+                logger->log_info("planner returns", timestep + timeout_timesteps);
+                return;
+            }
+        logger->log_info("planner timeout", timestep + timeout_timesteps);
+        timeout_timesteps += 1;
     }
-    logger->log_info("planner timeout", timestep);
-    return;
+
+    //
 }
 
 
@@ -141,24 +128,16 @@ void BaseSystem::log_preprocessing(bool succ)
     logger->flush();
 }
 
-// void BaseSystem::log_event_assigned(int agent_id, int task_id, int timestep)
-// {
-//     logger->log_info("Task " + std::to_string(task_id) + " is assigned to agent " + std::to_string(agent_id), timestep);
-// }
-
-
-// Moved to TaskManager
-// void BaseSystem::log_event_finished(int agent_id, int task_id, int timestep) 
-// {
-//     logger->log_info("Agent " + std::to_string(agent_id) + " finishes task " + std::to_string(task_id), timestep);
-// }
-
 
 void BaseSystem::simulate(int simulation_time)
 {
     //init logger
     //Logger* log = new Logger();
     initialize();
+
+    this->simulation_time = simulation_time;
+
+    vector<Action> all_wait_actions(num_of_agents, Action::NA);
 
     for (; simulator.get_curr_timestep() < simulation_time; )
     {
@@ -167,11 +146,31 @@ void BaseSystem::simulate(int simulation_time)
 
         auto start = std::chrono::steady_clock::now();
 
-        vector<Action> actions;
-        vector<int> proposed_schedule;
-        plan(actions,proposed_schedule);
+        int timeout_timesteps = 0;
+
+        plan(timeout_timesteps);
 
         auto end = std::chrono::steady_clock::now();
+
+        for (int i = 0 ; i< timeout_timesteps; i ++){
+            simulator.move(all_wait_actions);
+            for (int k = 0; k < num_of_agents; k++)
+
+            for (int a = 0; a < num_of_agents; a++)
+                {
+                    if (!env->goal_locations[a].empty())
+                        solution_costs[a]++;
+                }
+        }
+
+        total_timetous+=timeout_timesteps;
+
+        if (simulator.get_curr_timestep() >= simulation_time){
+
+            auto diff = end-start;
+            planner_times.push_back(std::chrono::duration<double>(diff).count());
+            break;
+        }
 
         for (int a = 0; a < num_of_agents; a++)
         {
@@ -180,25 +179,13 @@ void BaseSystem::simulate(int simulation_time)
         }
 
         // move drives
-        //move(actions);
-        vector<State> curr_states = simulator.move(actions);
+        vector<State> curr_states = simulator.move(proposed_actions);
         int timestep = simulator.get_curr_timestep();
         // agents do not move
 
-        for (int k = 0; k < num_of_agents; k++)
-        {
-            paths[k].push_back(curr_states[k]);
-            actual_movements[k].push_back(actions[k]);
-        }
-        if (!planner_movements[0].empty() && planner_movements[0].back() == Action::NA)
-        {
-            planner_times.back()+=plan_time_limit;  //add planning time to last record
-        }
-        else
-        {
-            auto diff = end-start;
-            planner_times.push_back(std::chrono::duration<double>(diff).count());
-        }
+
+        auto diff = end-start;
+        planner_times.push_back(std::chrono::duration<double>(diff).count());
 
         // update tasks
         task_manager.update_tasks(curr_states, proposed_schedule, simulator.get_curr_timestep());
@@ -228,17 +215,18 @@ void BaseSystem::initialize()
         _exit(124);
 
     // initialize_goal_locations();
-    task_manager.reveal_tasks(timestep);
+    task_manager.reveal_tasks(timestep); //this also intialize env->new_tasks
 
     sync_shared_env();
 
-    actual_movements.resize(num_of_agents);
-    planner_movements.resize(num_of_agents);
     solution_costs.resize(num_of_agents);
     for (int a = 0; a < num_of_agents; a++)
     {
         solution_costs[a] = 0;
     }
+
+    proposed_actions.resize(num_of_agents, Action::W);
+    proposed_schedule.resize(num_of_agents, -1);
 }
 
 
@@ -249,29 +237,30 @@ void BaseSystem::saveResults(const string &fileName, int screen) const
     js["actionModel"] = "MAPF_T";
     js["version"] = "2024 LoRR";
 
-    std::string feasible = fast_mover_feasible ? "Yes" : "No";
-    js["AllValid"] = feasible;
+    // std::string feasible = fast_mover_feasible ? "Yes" : "No";
+    // js["AllValid"] = feasible;
 
     js["teamSize"] = num_of_agents;
 
     js["numTaskFinished"] = task_manager.num_of_task_finish;
-    int sum_of_cost = 0;
     int makespan = 0;
     if (num_of_agents > 0)
     {
-        sum_of_cost = solution_costs[0];
         makespan = solution_costs[0];
         for (int a = 1; a < num_of_agents; a++)
         {
-            sum_of_cost += solution_costs[a];
             if (solution_costs[a] > makespan)
             {
                 makespan = solution_costs[a];
             }
         }
     }
-    js["sumOfCost"] = sum_of_cost;
     js["makespan"] = makespan;
+
+    js["numPlannerErrors"] = simulator.get_number_errors();
+    js["numScheduleErrors"] = task_manager.get_number_errors();
+
+    js["numEntryTimeouts"] = total_timetous;
 
     // Save start locations[x,y,orientation]
     if (screen <= 2)