+++ src/TaskScheduler.cpp
@@ -3,6 +3,16 @@
 #include "scheduler.h"
 #include "const.h"
 
+/**
+ * Initializes the task scheduler with a given time limit for preprocessing.
+ * 
+ * This function prepares the task scheduler by allocating up to half of the given preprocessing time limit 
+ * and adjust for a specified tolerance to account for potential timing errors. 
+ * It ensures that initialization does not exceed the allocated time.
+ * 
+ * @param preprocess_time_limit The total time limit allocated for preprocessing (in milliseconds).
+ *
+ */
 void TaskScheduler::initialize(int preprocess_time_limit)
 {
     //give at most half of the entry time_limit to scheduler;
@@ -10,6 +20,16 @@ void TaskScheduler::initialize(int preprocess_time_limit)
     int limit = preprocess_time_limit/2 - DefaultPlanner::SCHEDULER_TIMELIMIT_TOLERANCE;
     DefaultPlanner::schedule_initialize(limit, env);    
 }
+
+/**
+ * Plans a task schedule within a specified time limit.
+ * 
+ * This function schedules tasks by calling shedule_plan function in default planner with half of the given time limit,
+ * adjusted for timing error tolerance. The planned schedule is output to the provided schedule vector.
+ * 
+ * @param time_limit The total time limit allocated for scheduling (in milliseconds).
+ * @param proposed_schedule A reference to a vector that will be populated with the proposed schedule (next task id for each agent).
+ */
 
 void TaskScheduler::plan(int time_limit, std::vector<int> & proposed_schedule)
 {